AI/project/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
Pillow==10.2.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.1
django-ratelimit==4.1.0
